refactor: extract provider stream wrappers
